Add files via upload
This Python program implements a Caesar cipher, a simple encryption technique that shifts letters in the alphabet by a fixed number. The program can encrypt and decrypt text, preserving uppercase and lowercase letters, as well as spaces and punctuation.

In this version, I experimented with the text "Konka is the GOAT" to better understand the encryption and decryption process. After understanding how the cipher works, I modified it for fun and learning purposes.